Move schema/features.ts → shared/features.ts (web-shared foundation layer)
FeatureToggle and FeaturesResponse are a server↔viewer contract with two
distinct consumer zones, satisfying the web-shared two-consumer rule.
Co-locating with data-files.ts makes the shared foundation pattern explicit
and reduces the web-server→web zone crossing to shared-only constants
(eliminating the schema/ reverse import).

- Add features.ts to shared/ with updated docstring
- Export FeatureToggle/FeaturesResponse through shared/index.ts barrel
- Update server/routes-features.ts to import via ../shared/index.js
- Update viewer/external.ts to import from ../shared/features.js
- Add "features" to SHARED_LEAF_FILES in boundary-check.test.ts
- Delete packages/web/src/schema/features.ts
