+++ packages/web/src/shared/index.ts
@@ -8,3 +8,4 @@
 
 export { DATA_FILES, ALL_DATA_FILES, SUPPLEMENTARY_FILES } from "./data-files.js";
 export type { ViewId } from "./view-id.js";
+export type { FeatureToggle, FeaturesResponse } from "./features.js";
