+++ packages/web/src/viewer/external.ts
@@ -32,7 +32,7 @@ export type {
 export * as V1 from "../schema/v1.js";
 
 // ── Feature toggle contract types ──────────────────────────────────────────
-export type { FeatureToggle, FeaturesResponse } from "../schema/features.js";
+export type { FeatureToggle, FeaturesResponse } from "../shared/features.js";
 
 // ── Shared data-file constants ─────────────────────────────────────────────
 export { DATA_FILES, ALL_DATA_FILES, SUPPLEMENTARY_FILES } from "../shared/data-files.js";
