+++ packages/web/src/viewer/views/feature-toggles.ts
@@ -13,7 +13,7 @@ import { h } from "preact";
 import { useState, useEffect, useCallback, useRef } from "preact/hooks";
 import { NdxLogoPng } from "../components/logos.js";
 
-// ── Types (canonical definitions in src/schema/features.ts) ──────────
+// ── Types (canonical definitions in src/shared/features.ts) ──────────
 import type { FeatureToggle, FeaturesResponse } from "../external.js";
 
 // ── Package metadata ─────────────────────────────────────────────────
